move FM API wrapper to service
